Teams keep who they want, build around a core, and a traded pick finally means something
Four owner corrections in sequence, each of which found a real gap.

1. "a team traded away a 92 rated TE on an expiring contract (that they could have
   re-signed)"

⚠️ THE TRIGGER WAS TWO THIRDS BUILT. The plan states it as "walk-year, OVER THE RE-SIGN
LIMIT, not contending" and the middle clause was never implemented, so it fired on EVERY
expiring player and "he leaves for nothing" — the entire premise — was false most of the
time. Measured: a 1-14 club shipped a 93-rated kicker for one prospect at an ask of 0.3,
and a 7-10 club a 96-rated back. Both were the first name on their own re-sign list.

⚠️ And the live constraint is the PER-OFFSEASON cap, not a per-player one (owner: "theres
no longer a re-sign limit for players, only the amount of players that can be re-signed
every off season"). `RESIGN_ONCE_ENABLED` has been False since 2026-08-13.
`hasReachedResignLimit` is still consulted and is inert by design — it is the single source
of truth for re-sign eligibility, so the market reads it rather than growing a second
opinion that would need finding if that rule came back.

2. "teams need to be thinking about which players on expiring contracts they want to keep
   and which they deem expendable (this can still be highly rated players as long as the
   return is worth it)"

The cap says who is DEFINITELY leaving; it does not make the rest untouchable. A keepable
walk-year player is still listed, under `expiring_keeper`, priced on the contract that
would follow rather than on thirteen weeks of football — the same 92 asks 31x what he does
as surplus, so only a return that genuinely beats keeping him clears.

⚠️ THAT NEEDED ONE RULE, NOT TWO. Priced on the seller's side alone, the ask sat around 108
against a buyer still valuing a rental at 3.5, and NOT ONE keeper ever sold: 245 listed,
zero clearing bids. `retentionTerm` is now the single rule and both sides read it — a buyer
inherits the same right to re-sign him, THROUGH ITS OWN CAP. That is what makes the same
player genuinely expendable to one club and untouchable to another.

3. "teams should also be identifying star players to build around and not consider every
   highly rated player as a trade asset"

⚠️ BEING HIGHLY RATED IS NOT THE SAME AS BEING AVAILABLE. Every value trigger priced a star
as an asset with a big number on it, so a club's best man went on the block whenever the
arithmetic cleared — which is how a rebuilder sells the one player its rebuild is for. A
club's top two 4-star-or-better players are exempt.

⚠️ THE TWO GUARDS ON THAT RULE ARE THE LOAD-BEARING PART. "The best two on a 2-14 club"
would make the worst clubs untouchable and stop them trading at all, so it takes the game's
own star bar. And a DECLINING star stays an asset: selling high on a fading veteran is one
of the few genuinely smart things a front office can do, and the core is who you build
AROUND. `locker_room` still reaches a core player, because that is a reason to move someone
which has nothing to do with what he is worth.

4. "is the trade volume issue do to lack of available assets?"

⚠️ NO, AND THE MEASUREMENT SAYS SO PLAINLY. Of every bid attempt: 34% built a real bid, 30%
could not reach the seller's price, 34% the buyer refused because the cost exceeded the
gain — and only 1% held NOTHING to offer. Volume is the market declining, not a supply
shortage. The harness now reports that split, because "no bid" is indistinguishable
between an empty cupboard and a club that looked and passed, and those need opposite
responses.

AND A REAL BUG THAT QUESTION UNCOVERED:

⚠️ EVERY TRADED PICK WAS COSMETIC. `_runRookieDraftPhase` read `freeAgencyOrder` straight
through and never consulted `DraftPick` at all, so a club could trade for the first
selection, watch the transactions page say so, and then not get it. A quarter of every
bundle in the measured ledger was picks, all of it paying for nothing.
`_applyPickOwnership` re-points each slot at its owner and stamps it spent. ⚠️ It may
return the same club twice — a club holding two slots picks twice, and de-duplicating
would silently void the pick somebody paid for.

⚠️ `seasonsOfControl` also had to go float. A keeper's expected re-sign term is fractional
(a TierA averages 3.5 seasons) and `int()` truncates toward zero, which would quietly shave
most of a season off every keeper — the same silent-bias class `_rnd` exists to prevent in
the game engine.

MEASURED over six seasons, all four triggers firing, rosters intact (0 holes, 0 wrong-team
refs), in-season trades still entirely weeks 15-21. Volume runs ~0.5 trades per club per
season, well inside the plan's 2-3 cap.

⚠️ ONE METRIC IS STILL WRONG AND IS LABELLED AS SUCH: "upgrade taken" compares RATINGS, so
a trade that swaps similar players and gains three years of control reads as churn. The
harness now reports control gained alongside it.

Tests: 9 added across the four changes, every one shown to fail first — a dropped re-sign
cap, a keeper priced as a rental, a buyer ignoring its own cap, no core, a core that
ignores the star bar, a core that shields a declining player, a core that blocks the
locker-room trigger, a draft that ignores pick ownership, and one that de-duplicates an
owner out of his second selection.

⚠️ Several fixtures had to be rebuilt rather than adjusted: they gave a club ONE expiring
player, which under the corrected rule is simply re-signed, so they were testing a league
that cannot exist. `_congest` and `_withCore` make the fixture match the rule it names.

Suite: 219 pass, 2 fail — both pre-existing on clean development.

+++ trading.py
@@ -68,9 +68,14 @@ def seasonsOfControl(termRemaining: int, week: int = None) -> float:
     and everyone remaining has WHOLE seasons of term, so offseason trades are about
     ASSETS and the rental market does not exist there at all.
     """
-    term = max(0, int(termRemaining or 0))
+    # ⚠️ FLOAT, NOT INT. A walk-year player the club intends to re-sign is priced on the
+    # contract that would follow, and that expectation is fractional (a TierA averages 3.5
+    # seasons). `int()` truncates toward zero, so rounding here would quietly shave most of
+    # a season off every keeper's valuation — the same class of silent bias `_rnd` exists
+    # to prevent in the game engine.
+    term = max(0.0, float(termRemaining or 0))
     if week is None:
-        return float(term)
+        return term
     if term <= 0:
         return 0.0
     return (term - 1) + seasonRemainingFraction(week)
